+++ src/main.ts
@@ -1,5 +1,4 @@
 export * from './model-creator';
 export * from './mongoose-connection';
-export { ModelConfiguration } from './model-config';
 export { QueryWithIrrelevant } from './schema-helpers/query-with-irrelevant-wrapper';
 export { getCurrentDataVersion } from './data-version/data-version-manager';