+++ src/model-creator.ts
@@ -1,8 +1,11 @@
-import { PolarisBaseContext, PolarisRequestHeaders } from '@enigmatis/utills';
+import {
+    PolarisBaseContext,
+    PolarisRequestHeaders,
+    SoftDeleteConfiguration,
+} from '@enigmatis/utills';
 import * as Joi from 'joi';
 import { Model, model, models, Schema } from 'mongoose';
 import { addDataVersionMiddleware } from './data-version/data-version-middleware';
-import { ModelConfiguration } from './model-config';
 import { getCollectionName } from './schema-helpers/middleware-functions';
 import {
     addDocumentMiddleware,
@@ -32,9 +35,8 @@ export declare type SchemaCreator = (refNameCreator: (name: string) => string) =
 export const getModelCreator = <T>(
     collectionPrefix: string,
     schemaOrCreator: Schema | SchemaCreator,
-    modelConfiguration?: ModelConfiguration,
 ): ModelCreator<T> => {
-    return ({ headers }: PolarisBaseContext): Model<InnerModelType<T>> => {
+    return ({ headers, softDeleteConfiguration }: PolarisBaseContext): Model<InnerModelType<T>> => {
         const collectionName = getCollectionName(collectionPrefix, headers);
         return (
             models[collectionName] ||
@@ -44,7 +46,7 @@ export const getModelCreator = <T>(
                     collectionPrefix,
                     schemaOrCreator,
                     headers,
-                    modelConfiguration,
+                    softDeleteConfiguration,
                 ),
             )
         );
@@ -58,7 +60,7 @@ const createSchemaForModel = <T>(
     collectionPrefix: string,
     schemaOrCreator: Schema | SchemaCreator,
     headers: PolarisRequestHeaders,
-    modelConfiguration?: ModelConfiguration,
+    softDeleteConfiguration?: SoftDeleteConfiguration,
 ) => {
     const schema =
         schemaOrCreator instanceof Function
@@ -66,7 +68,7 @@ const createSchemaForModel = <T>(
             : schemaOrCreator.clone();
     headers = checkHeaders(headers);
     addFields(schema);
-    addQueryMiddleware(schema, headers, modelConfiguration);
+    addQueryMiddleware(schema, headers, softDeleteConfiguration);
     addDocumentMiddleware(schema, headers);
     addModelMiddleware(schema, headers);
     addDataVersionMiddleware(schema, headers);