+++ test/model-creator.test.ts
@@ -1,6 +1,9 @@
-import { PolarisBaseContext, PolarisRequestHeaders } from '@enigmatis/utills';
+import {
+    PolarisBaseContext,
+    PolarisRequestHeaders,
+    SoftDeleteConfiguration,
+} from '@enigmatis/utills';
 import { Schema } from 'mongoose';
-import { ModelConfiguration } from '../src/model-config';
 import { getModelCreator } from '../src/model-creator';
 import { deleted, notDeleted } from '../src/schema-helpers/constants';
 import * as MiddlewareFunctions from '../src/schema-helpers/middleware-functions';
@@ -157,14 +160,11 @@ describe('module creator', () => {
             expect(model.findOneAndRemove).toBe(MiddlewareFunctions.findOneAndSoftDelete);
         });
         test('soft delete not allowed', () => {
-            const softDeleteNotAllowedModelConfig: ModelConfiguration = {
+            const softDeleteNotAllowed: SoftDeleteConfiguration = {
                 allowSoftDelete: false,
             };
-            const modelCreator2 = getModelCreator(
-                'testing2',
-                personSchema,
-                softDeleteNotAllowedModelConfig,
-            );
+            context.softDeleteConfiguration = softDeleteNotAllowed;
+            const modelCreator2 = getModelCreator('testing2', personSchema);
             const model2 = modelCreator2(context);
             expect(model2.schema.statics).not.toEqual({
                 deleteMany: MiddlewareFunctions.softRemove,
@@ -192,27 +192,23 @@ describe('module creator', () => {
 
     describe("middleware's functions", () => {
         test('findHandlerFunc - soft delete return entities true', () => {
-            const softDeleteReturnEntitiesModelConfig: ModelConfiguration = {
+            const softDeleteReturnEntities: SoftDeleteConfiguration = {
                 softDeleteReturnEntities: true,
             };
-            const modelCreator2 = getModelCreator(
-                'testing2',
-                personSchema,
-                softDeleteReturnEntitiesModelConfig,
-            );
+            context.softDeleteConfiguration = softDeleteReturnEntities;
+            const modelCreator2 = getModelCreator('testing2', personSchema);
             const model2 = modelCreator2(context);
             const where = jest.fn();
             const conditions = { name: 'Dazdraperma' };
             const headers = { realityId: testReality };
             const findHandler = MiddlewareFunctions.getFindHandler(
                 headers,
-                softDeleteReturnEntitiesModelConfig,
+                softDeleteReturnEntities,
             );
             findHandler.call({ where, _conditions: conditions });
             expect(where).toHaveBeenCalledTimes(1);
             expect(where).toHaveBeenCalledWith(expect.not.objectContaining(notDeleted));
         });
-
         test('findHandlerFunc - add not deleted options to query', () => {
             const where = jest.fn();
             const conditions = { name: 'Dazdraperma' };